Add files via upload
Since not all esp32 modules (e. g. esp32u) have IO9 and IO10 outputs and lack of these outputs on DEV boards, 
I decided to redesign the board and adjust the program code. Now the IO9 and IO10 are replaced by IO27 and IO26 so it is possible provide tests on the DEV boards also.
The program code is adapted to the change of the board and has fixes for the stability of work.
Note: For the interface to work properly, it is necessary to connect it to a heat pump or a system that simulates the pump response. 
In the /piotools directory there is a simple program HP_Panasonic_Simul.ino that can be uploaded e. g. on Arduino Nano. 
Without communication from the heat pump, the interface resets after a few seconds.